+++ cmd/zed/lake/find/command.go
@@ -8,10 +8,10 @@ import (
 
 	"github.com/brimdata/zed/cli/outputflags"
 	zedlake "github.com/brimdata/zed/cmd/zed/lake"
-	"github.com/brimdata/zed/emitter"
 	"github.com/brimdata/zed/lake"
 	"github.com/brimdata/zed/lake/index"
 	"github.com/brimdata/zed/pkg/charm"
+	"github.com/brimdata/zed/zio/emitter"
 	"github.com/brimdata/zed/zng"
 	"github.com/brimdata/zed/zson"
 )
