+++ cmd/zed/lake/map/command.go
@@ -11,7 +11,6 @@ import (
 	zedlake "github.com/brimdata/zed/cmd/zed/lake"
 	"github.com/brimdata/zed/compiler"
 	"github.com/brimdata/zed/driver"
-	"github.com/brimdata/zed/emitter"
 	"github.com/brimdata/zed/lake"
 	"github.com/brimdata/zed/lake/chunk"
 	"github.com/brimdata/zed/pkg/charm"
@@ -21,6 +20,7 @@ import (
 	"github.com/brimdata/zed/zbuf"
 	"github.com/brimdata/zed/zio"
 	"github.com/brimdata/zed/zio/detector"
+	"github.com/brimdata/zed/zio/emitter"
 	"github.com/brimdata/zed/zson"
 )
