+++ cmd/zed/lake/stat/command.go
@@ -7,11 +7,11 @@ import (
 	"os"
 
 	zedlake "github.com/brimdata/zed/cmd/zed/lake"
-	"github.com/brimdata/zed/emitter"
 	"github.com/brimdata/zed/lake"
 	"github.com/brimdata/zed/pkg/charm"
 	"github.com/brimdata/zed/zbuf"
 	"github.com/brimdata/zed/zio"
+	"github.com/brimdata/zed/zio/emitter"
 	"github.com/brimdata/zed/zson"
 )
