+++ pkg/test/internal.go
@@ -8,10 +8,10 @@ import (
 
 	"github.com/brimdata/zed/compiler"
 	"github.com/brimdata/zed/driver"
-	"github.com/brimdata/zed/emitter"
 	"github.com/brimdata/zed/zbuf"
 	"github.com/brimdata/zed/zio"
 	"github.com/brimdata/zed/zio/detector"
+	"github.com/brimdata/zed/zio/emitter"
 	"github.com/brimdata/zed/zson"
 )
